+++ src/ExportJob/apiQuery.js
@@ -1,23 +1,21 @@
-export const useExportJobQuery = () => {
-  return {
-    isLoading: false,
-    exportJob: {
-      metadata: {
-        createdDate: '2020-08-12T12:18:21.123+00:00',
-        updatedDate: '2020-08-12T12:18:21.123+00:00',
-      },
-      id: 2,
-      jobId: '102',
-      status: 'Successful',
-      type: 'bursar',
-      description: '# of charges: 20',
-      source: 'System',
-      startTime: '2020-08-12T12:21:21.123+00:00',
-      endTime: '2020-08-12T12:25:21.123+00:00',
-      files: [
-        '/img/tenant-assets/opentown-libraries-logo.c96ff678691e1a345321b50941335d81.png',
-        '/img/tenant-assets/opentown-libraries-logo.c96ff678691e1a345321b50941335d81.png',
-      ],
+import { useQuery } from 'react-query';
+
+import { useOkapiKy } from '@folio/stripes/core';
+
+export const useExportJobQuery = (id) => {
+  const ky = useOkapiKy();
+
+  const { isLoading, data = {} } = useQuery({
+    queryKey: ['ui-export-manager', 'export-job', id],
+    queryFn: async () => {
+      const response = await ky.get(`data-export-spring/jobs/${id}`).json();
+
+      return response;
     },
+  });
+
+  return {
+    isLoading,
+    exportJob: data,
   };
 };