+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 * Export manager - export jobs view (results). Refs UIEXPMGR-3.
 * Export manager - export job details. Refs UIEXPMGR-5.
 * Export manager - export jobs search and filters. Refs UIEXPMGR-4.
+* Show export job real data. Refs UIEXPMGR-10.